backoffice/requirements.txt
Django==3.1.7
mysqlclient
django-mysql
mysqlclient
django-cors-headers
djangorestframework
pika
kafka-python